% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country-alpha-2.R
\name{country_alpha_2}
\alias{country_alpha_2}
\title{Get country name from two letter country code}
\usage{
country_alpha_2(alpha_2)
}
\arguments{
\item{alpha_2}{two letter country code}
}
\value{
country name
}
\description{
Get country name from two letter country code
}
\examples{
country_alpha_2("DE")
country_alpha_2("ZA")
}
