% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country-alpha-3.R
\name{country_alpha_3}
\alias{country_alpha_3}
\title{Get country name from three letter country code}
\usage{
country_alpha_3(alpha_3)
}
\arguments{
\item{alpha_3}{three letter country code}
}
\value{
country name
}
\description{
Get country name from three letter country code
}
\examples{
country_alpha_3("DEU")
country_alpha_3("ZAF")
}
