% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country-numeric.R
\name{country_numeric}
\alias{country_numeric}
\title{Get country name from numeric country code}
\usage{
country_numeric(num)
}
\arguments{
\item{num}{numeric country code}
}
\value{
country name
}
\description{
Get country name from numeric country code
}
\examples{
country_numeric(276)
country_numeric(710)
}
