% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-data.R
\docType{data}
\name{iso_3166}
\alias{iso_3166}
\title{Country Codes from the ISO 3166 standard}
\format{A data frame with 249 records and 4 fields:
\describe{
  \item{country}{country name}
  \item{alpha_2}{two letter code}
  \item{alpha_3}{three letter code}
  \item{num}{numeric code}
}}
\usage{
iso_3166
}
\description{
Country Codes from the ISO 3166 standard
}
\keyword{datasets}
